Split up data factory code into multiple files
